Add support for OTA updates for client/server apps that use WFB
Add support to allow client/server apps that use WFB protocol to initiate and do an OTA update of the client over the WFB radio link.
The actual OTA protocol and logic is handled by the client/server apps, not by the WFB stack. WFB just provides a way to exchange OTA messages, for clients and clients where one end might not be using WFB natively.
The WFB protocol update is transparent for the (existing/future) clients. No change is required in existing WFB clients.